移动的相关功能基本实现
修复了很多bug，
加入了rtt的测试

todo：延迟补偿，碰撞检测

+++ client/Assets/Scripts/GoButton.cs
@@ -24,7 +24,7 @@ public class GoButton : MonoBehaviour
     public void Go() {
         InfoKeeper.Instance.ip = ipIF.text;
         InfoKeeper.Instance.port = int.Parse(portIF.text);
-        InfoKeeper.Instance.tickRate = float.Parse(tickIF.text);
+        InfoKeeper.Instance.tickRate = int.Parse(tickIF.text);
         SceneManager.LoadScene(2);
     }
 }
